changeService

+++ src/BMIService.java
@@ -1,6 +1,6 @@
 public class BMIService {
-    public double calculate(double metre, double kilogram) {
+    public int calculate(double metre, double kilogram) {
         double index = kilogram / (metre * metre);
-        return index;
+        return (int)index;
     }
 }
