+++ src/Main.java
@@ -3,7 +3,7 @@ public class Main {
         BMIService service = new BMIService();
         double metre = 1.87; // рост
         double kilogram = 98; // вес
-        int index = (int) service.calculate(metre,kilogram);
+        int index = service.calculate(metre,kilogram);
 
         System.out.println(index);
     }
